SSCSSI-165 Remove feature toggle from AssociatedCaseLinkHelper
