+++ charts/sscs-tribunals-api/values.ccd.preview.template.yaml
@@ -42,7 +42,6 @@ java:
     DELETED_REDACTED_DOC_ENABLED: true
     DUMMY: true
     CITIZEN_LOGIN_SERVICE_V2: true
-    ADD_LINK_TO_OTHER_ASSOCIATED_CASES_V2: true
     TOPIC_NAME: "${SERVICE_NAME}-servicebus-main-topic"
     SUBSCRIPTION_NAME: "${SERVICE_NAME}-servicebus-main-topic"
     AMQP_HOST: "sscs-sb-preview.servicebus.windows.net"
