+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/AssociatedCaseLinkHelper.java
@@ -4,7 +4,6 @@ import java.util.*;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
 import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
@@ -19,8 +18,6 @@ public class AssociatedCaseLinkHelper {
     private final CcdService ccdService;
     private final IdamService idamService;
     private final UpdateCcdCaseService updateCcdCaseService;
-    @Value("${feature.add-link-to-other-associated-cases-v2.enabled}")
-    private boolean addLinkToOtherAssociatedCasesV2Enabled;
 
     @Autowired
     public AssociatedCaseLinkHelper(CcdService ccdService, IdamService idamService, UpdateCcdCaseService updateCcdCaseService) {
@@ -68,32 +65,12 @@ public class AssociatedCaseLinkHelper {
         if (!matchedByNinoCases.isEmpty()) {
             caseData.setAssociatedCase(new ArrayList<>(associatedCases));
             caseData.setLinkedCasesBoolean("Yes");
-            if (addLinkToOtherAssociatedCasesV2Enabled) {
-                addLinkToOtherAssociatedCasesV2Enabled(matchedByNinoCases, caseData.getCcdCaseId());
-            } else {
-                addLinkToOtherAssociatedCases(matchedByNinoCases, caseData.getCcdCaseId());
-            }
+            addLinkToOtherAssociatedCasesV2Enabled(matchedByNinoCases, caseData.getCcdCaseId());
 
         } else {
             caseData.setLinkedCasesBoolean("No");
         }
         return caseData;
-    }
-
-    private void addLinkToOtherAssociatedCases(List<SscsCaseDetails> matchedByNinoCases, String caseId) {
-        log.info("Adding link to other associated cases");
-        if (!matchedByNinoCases.isEmpty() && !StringUtils.isEmpty(caseId)) {
-            for (SscsCaseDetails sscsCaseDetails: matchedByNinoCases) {
-                SscsCaseData sscsCaseData = sscsCaseDetails.getData();
-                List<CaseLink> linkList = Optional.ofNullable(sscsCaseData.getAssociatedCase()).orElse(new ArrayList<>());
-                linkList.add(CaseLink.builder().value(
-                        CaseLinkDetails.builder().caseReference(caseId).build()).build());
-                Set<CaseLink> uniqueLinkSet = new HashSet<>(linkList);
-                sscsCaseData.setAssociatedCase(new ArrayList<>(uniqueLinkSet));
-                sscsCaseData.setLinkedCasesBoolean("Yes");
-                ccdService.updateCase(sscsCaseData, Long.valueOf(sscsCaseData.getCcdCaseId()), EventType.UPDATE_CASE_ONLY.getCcdType(), "updated case only", "Auto linked case added", idamService.getIdamTokens());
-            }
-        }
     }
 
     private void addLinkToOtherAssociatedCasesV2Enabled(List<SscsCaseDetails> matchedByNinoCases, String caseId) {
